refactor(agent): extract commit.py from graph.py
